+++ renren-crawler/src/main/java/io/renren/modules/crawler/ydzx/service/TbDetailsService.java
@@ -14,7 +14,7 @@ import java.util.Map;
  */
 public interface TbDetailsService {
 	
-	long queryMaxId();
+	int queryMaxId();
 	
 	TbDetailsEntity queryObject(Long id);
